+++ scilab/modules/gui/src/java/org/scilab/modules/gui/events/Jxgetmouse.java
@@ -19,13 +19,13 @@ public class Jxgetmouse {
 	/**
 	 * @return the xCoordinate
 	 */
-	public static float getXCoordinate() {
+	public static double getXCoordinate() {
 		return ClickInfos.getInstance().getXCoordinate();
 	}
 	/**
 	 * @return the yCoordinate
 	 */
-	public static float getYCoordinate() {
+	public static double getYCoordinate() {
 		return ClickInfos.getInstance().getXCoordinate();
 	}
 }
